organize clob metadata types
